+++ src/Http/Session.php
@@ -168,12 +168,7 @@ class Session
         if (Config::exists('App.sessionTimeout')) {
             $timeout = Config::read('App.sessionTimeout');
         }
-        /**
-         * @deprecated left for backwards comptability
-         */
-        if (Config::exists('Session.timeout')) {
-            $timeout = Config::read('Session.timeout');
-        }
+
         $lastActivity = $this->read('Session.lastActivity');
         $this->write('Session.lastActivity', time());