+++ src/Model/Schema/MysqlSchema.php
@@ -632,15 +632,7 @@ class MysqlSchema extends BaseSchema
         }
 
         # MySQL Options
-        if (isset($options['options'])) {
-            $databaseOptions = $options['options'];
-            /**
-             * @deprecated This provides backwards comptability
-             */
-            if (is_string($databaseOptions)) {
-                $databaseOptions = ['options' => $options];
-            }
-        }
+        $databaseOptions = $options['options'] ?? [];
 
         if (isset($options['options']['autoIncrement']) && isset($options['constraints']['primary']['column'])) {
             if (is_string($options['constraints']['primary']['column'])) {