+++ tests/TestApp/config/cache.php
@@ -1,3 +1,13 @@
 <?php
 
-return [];
+use Origin\Cache\Engine\FileEngine;
+
+return [
+    'origin' => [
+        'className' => FileEngine::class,
+        'path' => CACHE . '/origin',
+        'duration' => '+1 minutes',
+        'prefix' => 'cache_',
+        'serialize' => true
+    ]
+];